current_time save working

+++ db/migrate/20210407142441_create_user_episodes.rb
@@ -4,7 +4,7 @@ class CreateUserEpisodes < ActiveRecord::Migration[6.1]
       t.belongs_to :user, null: false, foreign_key: true
       t.integer :podcast_id
       t.boolean :listened, default: false
-      t.integer :current_time, default: 0
+      t.float :current_time, default: 0
       t.string :title
 
       t.timestamps
